[policies] Refactor util fn
