+++ app/javascript/src/Policies/reducers/PolicyRegistry.jsx
@@ -1,33 +1,10 @@
 // @flow
 
 import { initialState } from 'Policies/reducers/initialState'
-import { createReducer, updateArray } from 'Policies/reducers/util'
+import { createReducer, updateArray, parsePolicies } from 'Policies/util'
 
 import type { RegistryState } from 'Policies/types/State'
-import type { RawPolicy, RawRegistry, RegistryPolicy } from 'Policies/types/Policies'
 import type { FetchRegistrySuccessAction } from 'Policies/actions/PolicyRegistry'
-
-function parsePolicies (registry: RawRegistry): Array<RegistryPolicy> {
-  let policies: Array<RegistryPolicy> = []
-  for (let key in registry) {
-    registry[key].forEach(policy => policies.push(parsePolicy(key, policy)))
-  }
-  return policies
-}
-
-function parsePolicy (key: string, policy: RawPolicy): RegistryPolicy {
-  return Object.assign({},
-    {
-      $schema: policy.$schema,
-      name: key,
-      humanName: policy.name,
-      schema: policy.configuration,
-      version: policy.version,
-      summary: policy.summary,
-      description: policy.description,
-      configuration: {}
-    })
-}
 
 function updateRegistry (state: RegistryState, action: FetchRegistrySuccessAction): RegistryState {
   return updateArray(state, parsePolicies(action.payload))
