+++ app/javascript/src/Policies/util.jsx
@@ -2,6 +2,7 @@
 
 import type { UIState } from 'Policies/types/State'
 import type { FetchErrorAction, Reducer } from 'Policies/types/index'
+import type { RawPolicy, RawRegistry, RegistryPolicy } from 'Policies/types/Policies'
 
 function updateObject (oldObject: Object, newValues: Object): Object {
   return {...oldObject, ...newValues}
@@ -35,10 +36,34 @@ function generateGuid (): string {
   return s4() + s4() + '-' + s4() + '-' + s4() + '-' + s4() + '-' + s4() + s4() + s4()
 }
 
+function parsePolicies (registry: RawRegistry): Array<RegistryPolicy> {
+  let policies: Array<RegistryPolicy> = []
+  for (let key in registry) {
+    registry[key].forEach(policy => policies.push(parsePolicy(key, policy)))
+  }
+  return policies
+}
+
+function parsePolicy (key: string, policy: RawPolicy): RegistryPolicy {
+  return Object.assign({},
+    {
+      $schema: policy.$schema,
+      name: key,
+      humanName: policy.name,
+      schema: policy.configuration,
+      version: policy.version,
+      summary: policy.summary,
+      description: policy.description,
+      configuration: {}
+    })
+}
+
 export {
   updateObject,
   updateArray,
   createReducer,
   updateError,
-  generateGuid
+  generateGuid,
+  parsePolicies,
+  parsePolicy
 }
